@@ -1,8 +1,8 @@
 import 'package:cloud_firestore/cloud_firestore.dart';
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
-import 'package:flutter_icons/flutter_icons.dart';
-import 'package:modal_progress_hud/modal_progress_hud.dart';
+import 'package:ionicons/ionicons.dart';
+import 'package:loading_overlay/loading_overlay.dart';
 import 'package:provider/provider.dart';
 import 'package:social_media_app/components/custom_image.dart';
 import 'package:social_media_app/models/user.dart';
@@ -19,7 +19,7 @@ class _CreatePostState extends State<CreatePost> {
   @override
   Widget build(BuildContext context) {
     currentUserId() {
-      return firebaseAuth.currentUser.uid;
+      return firebaseAuth.currentUser!.uid;
     }
 
     PostsViewModel viewModel = Provider.of<PostsViewModel>(context);
@@ -28,20 +28,20 @@ class _CreatePostState extends State<CreatePost> {
         await viewModel.resetPost();
         return true;
       },
-      child: ModalProgressHUD(
+      child: LoadingOverlay(
         progressIndicator: circularProgress(context),
-        inAsyncCall: viewModel.loading,
+        isLoading: viewModel.loading,
         child: Scaffold(
           key: viewModel.scaffoldKey,
           appBar: AppBar(
             leading: IconButton(
-              icon: Icon(Feather.x),
+              icon: Icon(Ionicons.close_outline),
               onPressed: () {
                 viewModel.resetPost();
                 Navigator.pop(context);
               },
             ),
-            title: Text('FlutterSocial'.toUpperCase()),
+            title: Text('WOOBLE'.toUpperCase()),
             centerTitle: true,
             actions: [
               GestureDetector(
@@ -57,7 +57,7 @@ class _CreatePostState extends State<CreatePost> {
                     style: TextStyle(
                       fontWeight: FontWeight.bold,
                       fontSize: 15.0,
-                      color: Theme.of(context).accentColor,
+                      color: Theme.of(context).colorScheme.secondary,
                     ),
                   ),
                 ),
@@ -72,18 +72,20 @@ class _CreatePostState extends State<CreatePost> {
                 stream: usersRef.doc(currentUserId()).snapshots(),
                 builder: (context, AsyncSnapshot<DocumentSnapshot> snapshot) {
                   if (snapshot.hasData) {
-                    UserModel user = UserModel.fromJson(snapshot.data.data());
+                    UserModel user = UserModel.fromJson(
+                      snapshot.data!.data() as Map<String, dynamic>,
+                    );
                     return ListTile(
                       leading: CircleAvatar(
                         radius: 25.0,
-                        backgroundImage: NetworkImage(user?.photoUrl),
+                        backgroundImage: NetworkImage(user.photoUrl!),
                       ),
                       title: Text(
-                        user?.username,
+                        user.username!,
                         style: TextStyle(fontWeight: FontWeight.bold),
                       ),
                       subtitle: Text(
-                        user?.email,
+                        user.email!,
                       ),
                     );
                   }
@@ -101,7 +103,7 @@ class _CreatePostState extends State<CreatePost> {
                       Radius.circular(5.0),
                     ),
                     border: Border.all(
-                      color: Theme.of(context).accentColor,
+                      color: Theme.of(context).colorScheme.secondary,
                     ),
                   ),
                   child: viewModel.imgLink != null
@@ -116,12 +118,13 @@ class _CreatePostState extends State<CreatePost> {
                               child: Text(
                                 'Upload a Photo',
                                 style: TextStyle(
-                                  color: Theme.of(context).accentColor,
+                                  color:
+                                      Theme.of(context).colorScheme.secondary,
                                 ),
                               ),
                             )
                           : Image.file(
-                              viewModel.mediaUrl,
+                              viewModel.mediaUrl!,
                               width: MediaQuery.of(context).size.width,
                               height: MediaQuery.of(context).size.width - 30,
                               fit: BoxFit.cover,
@@ -175,7 +178,7 @@ class _CreatePostState extends State<CreatePost> {
                     size: 25.0,
                   ),
                   iconSize: 30.0,
-                  color: Theme.of(context).accentColor,
+                  color: Theme.of(context).colorScheme.secondary,
                   onPressed: () => viewModel.getLocation(),
                 ),
               ),
@@ -210,7 +213,7 @@ class _CreatePostState extends State<CreatePost> {
               ),
               Divider(),
               ListTile(
-                leading: Icon(Feather.camera),
+                leading: Icon(Ionicons.camera_outline),
                 title: Text('Camera'),
                 onTap: () {
                   Navigator.pop(context);
@@ -218,7 +221,7 @@ class _CreatePostState extends State<CreatePost> {
                 },
               ),
               ListTile(
-                leading: Icon(Feather.image),
+                leading: Icon(Ionicons.image),
                 title: Text('Gallery'),
                 onTap: () {
                   Navigator.pop(context);
