@@ -1,7 +1,7 @@
 import 'package:cloud_firestore/cloud_firestore.dart';
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
-import 'package:flutter_icons/flutter_icons.dart';
+import 'package:ionicons/ionicons.dart';
 
 //import 'package:like_button/like_button.dart';
 import 'package:social_media_app/models/post.dart';
@@ -14,7 +14,7 @@ import 'package:social_media_app/widgets/cached_image.dart';
 import 'package:timeago/timeago.dart' as timeago;
 
 class Posts extends StatefulWidget {
-  final PostModel post;
+  final PostModel? post;
 
   Posts({this.post});
 
@@ -26,7 +26,7 @@ class _PostsState extends State<Posts> {
   final DateTime timestamp = DateTime.now();
 
   currentUserId() {
-    return firebaseAuth.currentUser.uid;
+    return firebaseAuth.currentUser!.uid;
   }
 
   //UserModel user;
@@ -50,16 +50,16 @@ class _PostsState extends State<Posts> {
             Container(
               height: 320.0,
               width: MediaQuery.of(context).size.width - 18.0,
-              child: cachedNetworkImage(widget.post.mediaUrl),
+              child: cachedNetworkImage(widget.post!.mediaUrl!),
             ),
             Flexible(
               child: ListTile(
                 contentPadding:
                     EdgeInsets.symmetric(horizontal: 10.0, vertical: 0.0),
                 title: Text(
-                  widget.post.description == null
+                  widget.post!.description! == null
                       ? ""
-                      : widget.post.description,
+                      : widget.post!.description!,
                   overflow: TextOverflow.ellipsis,
                   style: TextStyle(fontWeight: FontWeight.w700, fontSize: 12.0),
                 ),
@@ -68,19 +68,19 @@ class _PostsState extends State<Posts> {
                   child: Row(
                     children: [
                       Text(
-                        timeago.format(widget.post.timestamp.toDate()),
+                        timeago.format(widget.post!.timestamp!.toDate()),
                       ),
                       SizedBox(width: 3.0),
                       StreamBuilder(
                         stream: likesRef
-                            .where('postId', isEqualTo: widget.post.postId)
+                            .where('postId', isEqualTo: widget.post!.postId)
                             .snapshots(),
                         builder:
                             (context, AsyncSnapshot<QuerySnapshot> snapshot) {
                           if (snapshot.hasData) {
-                            QuerySnapshot snap = snapshot.data;
+                            QuerySnapshot snap = snapshot.data!;
                             List<DocumentSnapshot> docs = snap.docs;
-                            return buildLikesCount(context, docs?.length ?? 0);
+                            return buildLikesCount(context, docs.length ?? 0);
                           } else {
                             return buildLikesCount(context, 0);
                           }
@@ -89,16 +89,16 @@ class _PostsState extends State<Posts> {
                       SizedBox(width: 5.0),
                       StreamBuilder(
                         stream: commentRef
-                            .doc(widget.post.postId)
+                            .doc(widget.post!.postId)
                             .collection("comments")
                             .snapshots(),
                         builder:
                             (context, AsyncSnapshot<QuerySnapshot> snapshot) {
                           if (snapshot.hasData) {
-                            QuerySnapshot snap = snapshot.data;
+                            QuerySnapshot snap = snapshot.data!;
                             List<DocumentSnapshot> docs = snap.docs;
                             return buildCommentsCount(
-                                context, docs?.length ?? 0);
+                                context, docs.length ?? 0);
                           } else {
                             return buildCommentsCount(context, 0);
                           }
@@ -156,42 +156,42 @@ class _PostsState extends State<Posts> {
   }
 
   Widget buildPostHeader() {
-    bool isMe = currentUserId() == widget.post.ownerId;
+    bool isMe = currentUserId() == widget.post!.ownerId;
     return ListTile(
-        contentPadding: EdgeInsets.symmetric(horizontal: 5.0),
-        leading: buildUserDp(),
-        title: Text(
-    widget.post.username,
-    style: TextStyle(fontWeight: FontWeight.bold),
-        ),
-        subtitle: Text(
-    widget.post.location == null ? 'Wooble' : widget.post.location,
-        ),
-        trailing: isMe
-      ? IconButton(
-          icon: Icon(Feather.more_horizontal),
-          onPressed: () => handleDelete(context),
-        )
-      : IconButton(
-          ///Feature coming soon
-          icon: Icon(CupertinoIcons.bookmark, size: 25.0),
-          onPressed: () {},
-        ),
-      );
-
+      contentPadding: EdgeInsets.symmetric(horizontal: 5.0),
+      leading: buildUserDp(),
+      title: Text(
+        widget.post!.username!,
+        style: TextStyle(fontWeight: FontWeight.bold),
+      ),
+      subtitle: Text(
+        widget.post!.location! == null ? 'Wooble' : widget.post!.location!,
+      ),
+      trailing: isMe
+          ? IconButton(
+              icon: Icon(Ionicons.ellipsis_horizontal_outline),
+              onPressed: () => handleDelete(context),
+            )
+          : IconButton(
+              ///Feature coming soon
+              icon: Icon(CupertinoIcons.bookmark, size: 25.0),
+              onPressed: () {},
+            ),
+    );
   }
 
   buildUserDp() {
     return StreamBuilder(
-      stream: usersRef.doc(widget.post.ownerId).snapshots(),
+      stream: usersRef.doc(widget.post!.ownerId).snapshots(),
       builder: (context, AsyncSnapshot<DocumentSnapshot> snapshot) {
         if (snapshot.hasData) {
-          UserModel user = UserModel.fromJson(snapshot.data.data());
+          UserModel user =
+              UserModel.fromJson(snapshot.data!.data() as Map<String, dynamic>);
           return GestureDetector(
-            onTap: () => showProfile(context, profileId: user?.id),
+            onTap: () => showProfile(context, profileId: user.id!),
             child: CircleAvatar(
               radius: 25.0,
-              backgroundImage: NetworkImage(user.photoUrl),
+              backgroundImage: NetworkImage(user.photoUrl!),
             ),
           );
         }
@@ -203,18 +203,18 @@ class _PostsState extends State<Posts> {
   buildLikeButton() {
     return StreamBuilder(
       stream: likesRef
-          .where('postId', isEqualTo: widget.post.postId)
+          .where('postId', isEqualTo: widget.post!.postId)
           .where('userId', isEqualTo: currentUserId())
           .snapshots(),
       builder: (context, AsyncSnapshot<QuerySnapshot> snapshot) {
         if (snapshot.hasData) {
-          List<QueryDocumentSnapshot> docs = snapshot?.data?.docs ?? [];
+          List<QueryDocumentSnapshot> docs = snapshot.data?.docs ?? [];
           return IconButton(
             onPressed: () {
               if (docs.isEmpty) {
                 likesRef.add({
                   'userId': currentUserId(),
-                  'postId': widget.post.postId,
+                  'postId': widget.post!.postId,
                   'dateCreated': Timestamp.now(),
                 });
                 addLikesToNotification();
@@ -239,37 +239,37 @@ class _PostsState extends State<Posts> {
   }
 
   addLikesToNotification() async {
-    bool isNotMe = currentUserId() != widget.post.ownerId;
+    bool isNotMe = currentUserId() != widget.post!.ownerId;
 
     if (isNotMe) {
       DocumentSnapshot doc = await usersRef.doc(currentUserId()).get();
-      user = UserModel.fromJson(doc.data());
+      user = UserModel.fromJson(doc.data() as Map<String, dynamic>);
       notificationRef
-          .doc(widget.post.ownerId)
+          .doc(widget.post!.ownerId)
           .collection('notifications')
-          .doc(widget.post.postId)
+          .doc(widget.post!.postId)
           .set({
         "type": "like",
-        "username": user.username,
+        "username": user!.username,
         "userId": currentUserId(),
-        "userDp": user.photoUrl,
-        "postId": widget.post.postId,
-        "mediaUrl": widget.post.mediaUrl,
+        "userDp": user!.photoUrl,
+        "postId": widget.post!.postId,
+        "mediaUrl": widget.post!.mediaUrl,
         "timestamp": timestamp,
       });
     }
   }
 
   removeLikeFromNotification() async {
-    bool isNotMe = currentUserId() != widget.post.ownerId;
+    bool isNotMe = currentUserId() != widget.post!.ownerId;
 
     if (isNotMe) {
       DocumentSnapshot doc = await usersRef.doc(currentUserId()).get();
-      user = UserModel.fromJson(doc.data());
+      user = UserModel.fromJson(doc.data() as Map<String, dynamic>);
       notificationRef
-          .doc(widget.post.ownerId)
+          .doc(widget.post!.ownerId)
           .collection('notifications')
-          .doc(widget.post.postId)
+          .doc(widget.post!.postId)
           .get()
           .then((doc) => {
                 if (doc.exists) {doc.reference.delete()}
@@ -284,7 +284,8 @@ class _PostsState extends State<Posts> {
         builder: (context) {
           return SimpleDialog(
             shape: RoundedRectangleBorder(
-                borderRadius: BorderRadius.circular(5.0)),
+              borderRadius: BorderRadius.circular(5.0),
+            ),
             children: [
               SimpleDialogOption(
                 onPressed: () {
@@ -307,13 +308,13 @@ class _PostsState extends State<Posts> {
 
 //you can only delete your own posts
   deletePost() async {
-    postRef.doc(widget.post.id).delete();
+    postRef.doc(widget.post!.id).delete();
 
 //delete notification associated with that given post
     QuerySnapshot notificationsSnap = await notificationRef
-        .doc(widget.post.ownerId)
+        .doc(widget.post!.ownerId!)
         .collection('notifications')
-        .where('postId', isEqualTo: widget.post.postId)
+        .where('postId', isEqualTo: widget.post!.postId)
         .get();
     notificationsSnap.docs.forEach((doc) {
       if (doc.exists) {
@@ -323,7 +324,7 @@ class _PostsState extends State<Posts> {
 
 //delete all the comments associated with that given post
     QuerySnapshot commentSnapshot =
-        await commentRef.doc(widget.post.postId).collection('comments').get();
+        await commentRef.doc(widget.post!.postId).collection('comments').get();
     commentSnapshot.docs.forEach((doc) {
       if (doc.exists) {
         doc.reference.delete();
@@ -331,7 +332,7 @@ class _PostsState extends State<Posts> {
     });
   }
 
-  showProfile(BuildContext context, {String profileId}) {
+  showProfile(BuildContext context, {String? profileId}) {
     Navigator.push(
       context,
       CupertinoPageRoute(
