@@ -1,5 +1,6 @@
 import 'package:firebase_auth/firebase_auth.dart';
 import 'package:flutter/material.dart';
+import 'package:google_fonts/google_fonts.dart';
 import 'package:provider/provider.dart';
 import 'package:social_media_app/components/life_cycle_event_handler.dart';
 import 'package:social_media_app/landing/landing_page.dart';
@@ -8,6 +9,7 @@ import 'package:social_media_app/services/user_service.dart';
 import 'package:social_media_app/utils/config.dart';
 import 'package:social_media_app/utils/constants.dart';
 import 'package:social_media_app/utils/providers.dart';
+import 'package:social_media_app/view_models/theme/theme_view_model.dart';
 
 void main() async {
   WidgetsFlutterBinding.ensureInitialized();
@@ -36,24 +38,35 @@ class _MyAppState extends State<MyApp> {
   Widget build(BuildContext context) {
     return MultiProvider(
       providers: providers,
-      child: Consumer<ThemeNotifier>(
-        builder: (context, ThemeNotifier notifier, child) {
+      child: Consumer<ThemeProvider>(
+        builder: (context, ThemeProvider notifier, Widget? child) {
           return MaterialApp(
             title: Constants.appName,
             debugShowCheckedModeBanner: false,
-            theme: notifier.dark ? Constants.darkTheme : Constants.lightTheme,
+            theme: themeData(
+              notifier.dark ? Constants.darkTheme : Constants.lightTheme,
+            ),
             home: StreamBuilder(
               stream: FirebaseAuth.instance.authStateChanges(),
-              builder: (BuildContext context, AsyncSnapshot<User> snapshot) {
+              builder: ((BuildContext context, snapshot) {
                 if (snapshot.hasData) {
                   return TabScreen();
                 } else
                   return Landing();
-              },
+              }),
             ),
           );
         },
       ),
     );
   }
+
+  ThemeData themeData(ThemeData theme) {
+    return theme.copyWith(
+      textTheme: GoogleFonts.nunitoTextTheme(
+        theme.textTheme,
+      ),
+    );
+  }
 }
+
