@@ -1,5 +1,5 @@
 import 'package:flutter/material.dart';
-import 'package:shared_preferences/shared_preferences.dart';
+import 'package:google_fonts/google_fonts.dart';
 
 class Constants {
   //App related strings
@@ -17,10 +17,8 @@ class Constants {
   static Color darkBG = Color(0xff2B2B2B);
 
   static ThemeData lightTheme = ThemeData(
-    fontFamily: 'Lato-Regular',
     backgroundColor: lightBG,
     primaryColor: lightPrimary,
-    accentColor: lightAccent,
     textSelectionTheme: TextSelectionThemeData(
       cursorColor: lightAccent,
     ),
@@ -30,24 +28,35 @@ class Constants {
       color: lightBG,
     ),
     appBarTheme: AppBarTheme(
-      elevation: 0,
-      textTheme: TextTheme(
-        headline6: TextStyle(
-          color: Colors.black,
-          fontSize: 20,
-          fontWeight: FontWeight.w600,
-          fontFamily: 'Lato-Regular',
-        ),
+      elevation: 0.0,
+      backgroundColor: lightBG,
+      iconTheme: const IconThemeData(color: Colors.black),
+      toolbarTextStyle: GoogleFonts.nunito(
+        color: Colors.black,
+        fontSize: 20.0,
+        fontWeight: FontWeight.w700,
       ),
+      titleTextStyle: GoogleFonts.nunito(
+        color: Colors.black,
+        fontSize: 20.0,
+        fontWeight: FontWeight.w700,
+      ),
+    ),
+    colorScheme: ColorScheme.fromSwatch().copyWith(
+      secondary: lightAccent,
     ),
   );
 
   static ThemeData darkTheme = ThemeData(
-    fontFamily: 'Lato-Regular',
-    brightness: Brightness.dark,
+    iconTheme: const IconThemeData(color: Colors.white),
+    colorScheme: ColorScheme.fromSwatch(
+      accentColor: darkAccent,
+    ).copyWith(
+      secondary: darkAccent,
+      brightness: Brightness.dark,
+    ),
     backgroundColor: darkBG,
     primaryColor: darkPrimary,
-    accentColor: darkAccent,
     scaffoldBackgroundColor: darkBG,
     textSelectionTheme: TextSelectionThemeData(
       cursorColor: darkAccent,
@@ -57,14 +66,18 @@ class Constants {
       color: darkBG,
     ),
     appBarTheme: AppBarTheme(
-      elevation: 0,
-      textTheme: TextTheme(
-        headline6: TextStyle(
-          color: lightBG,
-          fontSize: 20,
-          fontWeight: FontWeight.w600,
-          fontFamily: 'Lato-Regular',
-        ),
+      elevation: 0.0,
+      backgroundColor: darkBG,
+      iconTheme: const IconThemeData(color: Colors.white),
+      toolbarTextStyle: GoogleFonts.nunito(
+        color: Colors.white,
+        fontSize: 20.0,
+        fontWeight: FontWeight.w700,
+      ),
+      titleTextStyle: GoogleFonts.nunito(
+        color: Colors.white,
+        fontSize: 20.0,
+        fontWeight: FontWeight.w700,
       ),
     ),
   );
@@ -76,37 +89,5 @@ class Constants {
     }
 
     return result;
-  }
-}
-
-class ThemeNotifier extends ChangeNotifier {
-  final String key = 'theme';
-  SharedPreferences _prefs;
-  bool _darkTheme;
-  bool get dark => _darkTheme;
-
-  ThemeNotifier() {
-    _darkTheme = true;
-    _loadfromPrefs();
-  }
-  toggleTheme() {
-    _darkTheme = !_darkTheme;
-    _saveToPrefs();
-    notifyListeners();
-  }
-
-  _initPrefs() async {
-    if (_prefs == null) _prefs = await SharedPreferences.getInstance();
-  }
-
-  _loadfromPrefs() async {
-    await _initPrefs();
-    _darkTheme = _prefs.getBool(key) ?? true;
-    notifyListeners();
-  }
-
-  _saveToPrefs() async {
-    await _initPrefs();
-    _prefs.setBool(key, _darkTheme);
   }
 }
